pyproject.toml
[tool.poetry]
name = "veldigitizer"
version = "0.1.0"
description = ""
authors = ["Kauri Kolehmainen <kaurikolehmainen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^9.5.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
